backend/requirements_backend.txt
pandas
openpyxl
mysql-connector-python
sqlalchemy
fastapi
uvicorn[standard]
python-multipart
alembic
passlib[bcrypt]
python-jose
websockets


python-dotenv
pymysql

